Moved all user commands to a separate cog
