+++ main.py
@@ -3,22 +3,20 @@ import concurrent.futures
 import logging
 import os
 from collections import defaultdict, deque
-from typing import Optional, Sequence
+from typing import Optional
 
 import discord
 from alembic import command as alembic_command
 from alembic.config import Config as AlembicConfig
-from discord import app_commands, Interaction, Object, Member, Embed, Colour
+from discord import app_commands, Interaction, Object, Embed, Colour
 from discord.ext.commands import ExtensionNotLoaded, AutoShardedBot, Cog
 from dotenv import load_dotenv
 from requests_futures.sessions import FuturesSession
 from sqlalchemy import CursorResult, delete, exists, func, insert, select, update
-from sqlalchemy.engine.row import Row
 from sqlalchemy.ext.asyncio import AsyncConnection, create_async_engine, AsyncEngine
-from sqlalchemy.sql.functions import count
 
 from consts import *
-from model import (BlacklistModel, Member, MemberModel, ServerConfig, ServerConfigModel, UsedWordsModel, WhitelistModel,
+from model import (BlacklistModel, MemberModel, ServerConfig, ServerConfigModel, UsedWordsModel, WhitelistModel,
                    WordCacheModel)
 from utils import calculate_total_karma, db_connection
 
@@ -736,7 +734,7 @@ The above entered word is **NOT** being taken into account.''')
 word_chain_bot: WordChainBot = WordChainBot()
 
 
-# ---------------------------------------------------------------------------------------------------------------
+# ===================================================================================================================
 
 
 @word_chain_bot.tree.command(name='reload', description='Unload and reload a cog')
@@ -746,6 +744,7 @@ word_chain_bot: WordChainBot = WordChainBot()
 @app_commands.choices(cog_name=[
     app_commands.Choice(name='Admin Commands', value=COG_NAME_ADMIN_CMDS),
     app_commands.Choice(name='Manager Commands', value=COG_NAME_MANAGER_CMDS),
+    app_commands.Choice(name='User Commands', value=COG_NAME_USER_CMDS),
     app_commands.Choice(name='All cogs', value='all')
 ])
 async def reload(interaction: Interaction, cog_name: str):
@@ -797,320 +796,8 @@ async def reload(interaction: Interaction, cog_name: str):
 
     await interaction.followup.send(embed=emb)
 
-# ---------------------------------------------------------------------------------------------------------------
-
-
-@word_chain_bot.tree.command(name='list_commands', description='List all slash commands')
-@app_commands.describe(ephemeral="Whether the list will be publicly displayed")
-async def list_commands(interaction: discord.Interaction, ephemeral: bool = True):
-    """Command to list all the slash commands"""
-
-    await interaction.response.defer()
-
-    emb = discord.Embed(title='Slash Commands', color=discord.Color.blue(),
-                        description='''
-**list_commands** - Lists all the slash commands.
-**stats user** - Shows the stats of a specific user.
-**stats server** - Shows the stats of the server.
-**check_word** - Check if a word exists/check the spelling.
-**leaderboard** - Shows the leaderboard of the server.''')
-
-    if interaction.user.guild_permissions.ban_members:
-        emb.description += '''\n
-__Restricted commands__ (Admin-only)
-**sync** - Syncs the slash commands to the bot.
-**set_channel** - Sets the channel to chain words.
-**set_failed_role** - Sets the role to give when a user fails.
-**set_reliable_role** - Sets the reliable role.
-**remove_failed_role** - Unsets the role to give when a user fails.
-**remove_reliable_role** - Unset the reliable role.
-**prune** - Remove data for users who are no longer in the server.
-**blacklist add** - Add a word to the blacklist for this server.
-**blacklist remove** - Remove a word from the blacklist of this server.
-**blacklist show** - Show the blacklisted words for this server.
-**whitelist add** - Add a word to the whitelist for this server.
-**whitelist remove** - Remove a word from the whitelist of this server.
-**whitelist show** - Show the whitelist words for this server.'''
-
-    await interaction.followup.send(embed=emb, ephemeral=ephemeral)
-
-# ---------------------------------------------------------------------------------------------------------------
-
-
-@word_chain_bot.tree.command(name='check_word', description='Check if a word is correct')
-@app_commands.describe(word='The word to check')
-async def check_word(interaction: discord.Interaction, word: str):
-    """
-    Checks if a word is valid.
-
-    Hierarchy followed:
-    1. Legal characters.
-    2. Length of word must be > 1.
-    3. Whitelist.
-    4. Blacklists
-    5. Check word cache.
-    6. Query API.
-    """
-    await interaction.response.defer()
-
-    emb = discord.Embed(color=discord.Color.blurple())
-
-    if not all(c in POSSIBLE_CHARACTERS for c in word.lower()):
-        emb.description = f'❌ **{word}** is **not** a legal word.'
-        await interaction.followup.send(embed=emb)
-        return
-
-    if len(word) == 1:
-        emb.description = f'❌ **{word}** is **not** a valid word.'
-        await interaction.followup.send(embed=emb)
-        return
-
-    word = word.lower()
-
-    async with db_connection(word_chain_bot) as connection:
-        if await word_chain_bot.is_word_whitelisted(word, interaction.guild.id, connection):
-            emb.description = f'✅ The word **{word}** is valid.'
-            await interaction.followup.send(embed=emb)
-            return
-
-        if await word_chain_bot.is_word_blacklisted(word, interaction.guild.id, connection):
-            emb.description = f'❌ The word **{word}** is **blacklisted** and hence, **not** valid.'
-            await interaction.followup.send(embed=emb)
-            return
-
-        if await word_chain_bot.is_word_in_cache(word, connection):
-            emb.description = f'✅ The word **{word}** is valid.'
-            await interaction.followup.send(embed=emb)
-            return
-
-        future: concurrent.futures.Future = word_chain_bot.start_api_query(word)
-
-        match word_chain_bot.get_query_response(future):
-            case word_chain_bot.API_RESPONSE_WORD_EXISTS:
-
-                emb.description = f'✅ The word **{word}** is valid.'
-
-                await word_chain_bot.add_to_cache(word, connection)
-
-            case word_chain_bot.API_RESPONSE_WORD_DOESNT_EXIST:
-                emb.description = f'❌ **{word}** is **not** a valid word.'
-            case _:
-                emb.description = f'⚠️ There was an issue in fetching the result.'
-
-        await interaction.followup.send(embed=emb)
 
 # ===================================================================================================================
-
-
-class LeaderboardCmdGroup(app_commands.Group):
-
-    def __init__(self):
-        super().__init__(name='leaderboard')
-
-    # ---------------------------------------------------------------------------------------------------------------
-
-    @app_commands.command(description='Shows the first 10 users with the highest score/karma')
-    @app_commands.describe(metric='Use either score or karma for ordering the leaderboard')
-    @app_commands.choices(metric=[
-        app_commands.Choice(name='score', value='score'),
-        app_commands.Choice(name='karma', value='karma')
-    ])
-    @app_commands.describe(scope='Use either users from the current server or all users globally for the leaderboard')
-    @app_commands.choices(scope=[
-        app_commands.Choice(name='server', value='server'),
-        app_commands.Choice(name='global', value='global')
-    ])
-    async def user(self, interaction: discord.Interaction, metric: Optional[app_commands.Choice[str]],
-                   scope: Optional[app_commands.Choice[str]]):
-        """Command to show the top 10 users with the highest score/karma."""
-        await interaction.response.defer()
-
-        board_metric: str = 'score' if metric is None else metric.value
-        board_scope: str = 'server' if scope is None else scope.value
-
-        emb = discord.Embed(
-            title=f'Top 10 users by {board_metric}',
-            color=discord.Color.blue(),
-            description=''
-        )
-
-        match board_scope:
-            case 'server':
-                emb.set_author(name=interaction.guild.name,
-                               icon_url=interaction.guild.icon.url if interaction.guild.icon else None)
-            case 'global':
-                emb.set_author(name='Global')
-
-        async with db_connection(word_chain_bot, locked=False) as connection:
-            limit = 10
-
-            match board_metric:
-                case 'score':
-                    field = MemberModel.score
-                case 'karma':
-                    field = MemberModel.karma
-                case _:
-                    raise ValueError(f'Unknown metric {board_metric}')
-
-            match board_scope:
-                case 'server':
-                    stmt = (select(MemberModel.member_id, field)
-                            .where(MemberModel.server_id == interaction.guild.id)
-                            .order_by(field.desc())
-                            .limit(limit))
-                case 'global':
-                    stmt = (select(MemberModel.member_id, func.sum(field))
-                            .group_by(MemberModel.member_id)
-                            .order_by(func.sum(field).desc())
-                            .limit(limit))
-                case _:
-                    raise ValueError(f'Unknown scope {board_scope}')
-
-            result: CursorResult = await connection.execute(stmt)
-            data: Sequence[Row[tuple[int, int | float]]] = result.fetchall()
-
-            if len(data) == 0:  # Stop when no users could be retrieved.
-                match board_scope:
-                    case 'server':
-                        emb.description = ':warning: No users have played in this server yet!'
-                    case 'global':
-                        emb.description = ':warning: No users have played yet!'
-            else:
-                for i, user_data in enumerate(data, 1):
-                    member_id, score_or_karma = user_data
-                    match board_metric:
-                        case 'score':
-                            emb.description += f'{i}. <@{member_id}> **{score_or_karma}**\n'
-                        case 'karma':
-                            emb.description += f'{i}. <@{member_id}> **{score_or_karma:.2f}**\n'
-
-            await interaction.followup.send(embed=emb)
-
-    # ---------------------------------------------------------------------------------------------------------------
-
-    @app_commands.command(description='Shows the first 10 servers with the highest highscore')
-    async def server(self, interaction: discord.Interaction):
-        """Command to show the top 10 servers with the highest highscore"""
-        await interaction.response.defer()
-
-        emb = discord.Embed(
-            title=f'Top 10 servers by highscore',
-            color=discord.Color.blue(),
-            description=''
-        ).set_author(name='Global')
-
-        async with db_connection(word_chain_bot, locked=False) as connection:
-            limit = 10
-
-            stmt = (select(ServerConfigModel.server_id, ServerConfigModel.high_score)
-                    .order_by(ServerConfigModel.high_score.desc())
-                    .limit(limit))
-
-            result: CursorResult = await connection.execute(stmt)
-            data: Sequence[Row[tuple[int, int]]] = result.fetchall()
-
-            guild_names = defaultdict(lambda: 'unknown', {g.id: g.name for g in word_chain_bot.guilds})
-            for i, server_data in enumerate(data, 1):
-                server_id, high_score = server_data
-                emb.description += f'{i}. {guild_names[server_id]} **{high_score}**\n'
-
-            await interaction.followup.send(embed=emb)
-
-# ===================================================================================================================
-
-
-class StatsCmdGroup(app_commands.Group):
-
-    def __init__(self):
-        super().__init__(name='stats')
-
-    # ---------------------------------------------------------------------------------------------------------------
-
-    @app_commands.command(description='Show the server stats for the word chain game')
-    async def server(self, interaction: discord.Interaction) -> None:
-        """Command to show the stats of the server"""
-        await interaction.response.defer()
-
-        config: ServerConfig = word_chain_bot.server_configs[interaction.guild.id]
-
-        if config.channel_id is None:  # channel not set yet
-            await interaction.followup.send("Counting channel not set yet!")
-            return
-
-        server_stats_embed = discord.Embed(
-            description=f'''Current Chain Length: {config.current_count}
-Longest chain length: {config.high_score}
-{f"**Last word:** {config.current_word}" if config.current_word else ""}
-{f"Last word by: <@{config.last_member_id}>" if config.last_member_id else ""}''',
-            color=discord.Color.blurple()
-        )
-        server_stats_embed.set_author(name=interaction.guild, icon_url=interaction.guild.icon if interaction.guild.icon else None)
-
-        await interaction.followup.send(embed=server_stats_embed)
-
-    # ---------------------------------------------------------------------------------------------------------------
-
-    @app_commands.command(description='Show the user stats for the word chain game')
-    @app_commands.describe(member="The user whose stats you want to see")
-    async def user(self, interaction: discord.Interaction, member: Optional[discord.Member]) -> None:
-        """Command to show the stats of a specific user"""
-        await interaction.response.defer()
-
-        if member is None:
-            member = interaction.user
-
-        def get_member_avatar() -> Optional[discord.Asset]:
-            if member.avatar:
-                return member.avatar
-            elif member.display_avatar:
-                return member.display_avatar
-            else:
-                return None
-
-        async with db_connection(word_chain_bot, locked=False) as connection:
-            stmt = select(MemberModel).where(
-                MemberModel.server_id == member.guild.id,
-                MemberModel.member_id == member.id
-            )
-            result: CursorResult = await connection.execute(stmt)
-            row = result.fetchone()
-
-            if row is None:
-                await interaction.followup.send('You have never played in this server!')
-                return
-
-            db_member = Member.model_validate(row)
-
-            stmt = select(count(MemberModel.member_id)).where(
-                MemberModel.server_id == member.guild.id,
-                MemberModel.score >= db_member.score
-            )
-            result: CursorResult = await connection.execute(stmt)
-            pos_by_score = result.scalar()
-
-            stmt = select(count(MemberModel.member_id)).where(
-                MemberModel.server_id == member.guild.id,
-                MemberModel.karma >= db_member.karma
-            )
-            result: CursorResult = await connection.execute(stmt)
-            pos_by_karma = result.scalar()
-
-            emb = discord.Embed(
-                color=discord.Color.blue(),
-                description=f'''**Score:** {db_member.score} (#{pos_by_score})
-**🌟Karma:** {db_member.karma:.2f} (#{pos_by_karma})
-**✅Correct:** {db_member.correct}
-**❌Wrong:** {db_member.wrong}
-**Accuracy:** {(db_member.correct / (db_member.correct + db_member.wrong)):.2%}'''
-            ).set_author(name=f"{member} | stats", icon_url=get_member_avatar())
-
-            await interaction.followup.send(embed=emb)
-
-
-# ===================================================================================================================
-
 
 if __name__ == '__main__':
-    word_chain_bot.tree.add_command(LeaderboardCmdGroup())
-    word_chain_bot.tree.add_command(StatsCmdGroup())
     word_chain_bot.run(os.getenv('TOKEN'), log_handler=None)
